fix: Remove `this` from several group contructors
On a layout `->addWidget()` and `->setLayout()` automatically assigns the parent, as such we don't need to keep doing this. Even Qt complains about it in the console.

Doing `this` in each constructor tells Qt "Hey, add us to the main layout!" which spams the console with warnings that it's trying to add a layout to the main class, despite it already having one.